Update bindings directory structure
