+++ model_api/cpp/py_bindings/py_vision_api.cpp
@@ -1,0 +1,17 @@
+/*
+ * Copyright (C) 2025 Intel Corporation
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <nanobind/nanobind.h>
+
+namespace nb = nanobind;
+
+void init_classification(nb::module_& m);
+void init_base_modules(nb::module_& m);
+
+NB_MODULE(py_model_api, m) {
+    m.doc() = "Nanobind binding for OpenVINO Vision API library";
+    init_base_modules(m);
+    init_classification(m);
+}
